Adding movie detail fragment and requirements to fetch current movie

+++ app/src/main/java/com/lucas/schiavini/movieapp/view/MovieListAdapter.kt
@@ -44,7 +44,7 @@ class MovieListAdapter(
     override fun onMovieClicked(v: View) {
         val id = v.movieId.text.toString()
         val action = MoviesListFragmentDirections.actionMovieFragmentToMovieDetail()
-        action.movieId = id
+        action.movieId = id.toInt()
         Navigation.findNavController(v).navigate(action)
     }
